+++ apps/public/src/app/(general)/bedankt/page.tsx
@@ -1,6 +1,11 @@
+import type { Metadata } from 'next'
 import { cookies } from 'next/headers'
 
 import { Bedankt } from './Bedankt'
+
+export const metadata: Metadata = {
+  title: 'Bedankt - Gemeente Amsterdam',
+}
 
 export default async () => {
   const cookieStore = await cookies()
