+++ apps/public/src/app/(general)/locatie/page.tsx
@@ -1,6 +1,11 @@
+import type { Metadata } from 'next'
 import { cookies } from 'next/headers'
 
 import { Locatie } from './Locatie'
+
+export const metadata: Metadata = {
+  title: 'Stap 1 van 4 - Beschrijf uw melding - Gemeente Amsterdam',
+}
 
 export default async () => {
   const cookieStore = await cookies()
